// n-backend/apps/api/product/src/core/detail/interfaces/mule-api-detail.interface.ts
export interface MuleProductDetail {
  code: string;
  name: string;
  mainCategory: string;
  departmentCode: string;
  lineCode: string;
  classCode: string;
  description: string;
  applicableStartDate: string;
  salesStartDate: string;
  salesEndDate: string;
  salesEndFlag: string;
  consumptionTaxRate: number;
  consumptionTaxCategory: string;
  designatedPharmaceuticalCategory?: string;
  liquorCategory?: string;
  tobaccoCategory?: string;
  hazardousMaterialFlag?: string;
  medicalEquipmentClass?: string;
  poisonousDeleteriousSubstanceCategory?: string;
  poisonousDrugCategory?: string;
  agriculturalChemicalCategory?: string;
  animalDrugCategory?: string;
  warrantyIssuanceCategory?: string;
  pharmaceuticalFlag?: string;
  salesTypeCategory?: string;
  warrantyCategory?: string;
  warrantyPeriod?: string;
  deliveryChargeCategoryEc?: string;
  onlineFlagEc?: string;
  onlineStartTimeEc?: string;
  onlineEndTimeEc?: string;
  consumptionTaxCategoryEc: string;
  onlineSalesCategoryEc: string;
  individualDeliveryChargeEc: number;
  setProductFlag: string;
  configurationFlag: string;
  customizedProductCategory: string;
  minOrderQuantity?: number;
  maxOrderQuantity?: number;
  stepQuantity?: number;
  storeMinOrderQuantity: number;
  storeMaxOrderQuantity: number;
  storeStepOrderQuantity: number;
  newPeriodFrom?: string;
  reserveFromStoreDisabled: string;
  sendToStoreDisabled: string;
  personalDeliveryDisabled: string;
  estimatedArrivalDays?: number;
  productWithServiceFlag: string;
  dropShippingCategory?: string;
  largeSizedProductCategoryEc?: string;
  originCountryCode?: string;
  flavorExpirationCategory?: string;
  flavorExpiration?: number;
  productCharacteristics?: string;
  colorPattern?: string;
  size?: string;
  brandName?: string;
  singleItemSizeWidth?: number;
  singleItemSizeHeight?: number;
  singleItemSizeDepth?: number;
  contentAmount?: number;
  contentAmountUnit?: string;
  modelNo?: string;
  lawsStandards?: string;
  singleItemSize2Width?: number;
  singleItemSize2Height?: number;
  singleItemSize2Depth?: number;
  specCategoryName01?: string;
  specCategoryValue01?: string;
  specCategoryName02?: string;
  specCategoryValue02?: string;
  specCategoryName03?: string;
  specCategoryValue03?: string;
  specCategoryName04?: string;
  specCategoryValue04?: string;
  specCategoryName05?: string;
  specCategoryValue05?: string;
  specCategoryName06?: string;
  specCategoryValue06?: string;
  specCategoryName07?: string;
  specCategoryValue07?: string;
  specCategoryName08?: string;
  specCategoryValue08?: string;
  specCategoryName09?: string;
  specCategoryValue09?: string;
  specCategoryName10?: string;
  specCategoryValue10?: string;
  specCategoryName11?: string;
  specCategoryValue11?: string;
  specCategoryName12?: string;
  specCategoryValue12?: string;
  specCategoryName13?: string;
  specCategoryValue13?: string;
  specCategoryName14?: string;
  specCategoryValue14?: string;
  specCategoryName15?: string;
  specCategoryValue15?: string;
  characterCopyrightNameEc?: string;
  longDescription?: string;
  cautions?: string;
  productSpecGroupId?: string;
  similarProduct?: string;
  recommendedProduct?: string;
  comparativeProduct?: string;
  kukuru?: string;
  variation01?: string;
  variation02?: string;
  variation03?: string;
  variation04?: string;
  variation05?: string;
  variationName01?: string;
  variationName02?: string;
  variationName03?: string;
  variationName04?: string;
  variationName05?: string;
  faceProductId?: string | null;
  serviceContents?: string | null;
}